cloud_functions/evernote-parser/requirements.txt
evernote3
oauth2
oauthlib
requests
requests_oauthlib
google-cloud-pubsub